requirements.txt
streamlit
tensorflow
opencv-python-headless
numpy
pillow
h5py